requirements.txt
pandas~=1.5.1
scikit-learn~=1.1.3
hydra-core==1.1.2
hydra-colorlog==1.1.0
mlflow~=1.30.0